grub.cfg - Search QEMU root device by label instead of hda/sda/vda paths
When the underlying device type for QEMU emulation changes from
IDE to VIRTIO to SCSI_VIRTIO the root device changes from /dev/hda2
to /dev/vda2 to /dev/sda2. This results in the need for manually
changing the kernels command line root parameter. Setting this to
the initial rootA partition label ( LABEL=resin-rootA ) mitigates this.

Signed-off-by: Praneeth Bodduluri <lifeeth@gmail.com>