Change test app name
A workaround for a problem with creating an application with CLI.
qemu86 app seems to be registered as a ppublic one, and older CLI versions
fail to create app with such a name even though app name has to be unique
in the scope of the user only. So the workaround is to change the app name
because we cannot update CLI version atm.

Change-type: patch
Changelog-entry: Fix automated balenaOS test run on QEMU
Signed-off-by: Roman Mazur <roman@balena.io>