docker-disk: Increase size of the data partition
By default the data partition is only 192M but on a non-emulated device
it is expanded to the whole of the partition. On an emulated device it
does not expand so we need to set a bigger default so the device is usable.

Fixes #158

Changelog-entry: Increase size of the data partition
Signed-off-by: Alex Gonzalez <alexg@balena.io>